Modernize program.cs
